# com.sap.cloud.lm.sl.cf.api/test_rest-mtarest-1.0.0-swagger.yaml
...
swagger: "2.0"
info:
  description: "This is an API for the deploy-service(MTA) application"
  version: "1.0.0"
  title: "MTA REST API"
  contact:
    email: "DL_563CD7577BCF844DBD000037@exchange.sap.corp"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
# host: "deploy-service.bosh-lite.com"
# basePath: "/api/v1/spaces/{space_guid}"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- oauth2: []
paths:
...
  /operations:
    get:
      description: "Retrieves Multi-Target Application operations\n"
      operationId: "GetMtaOperations"
      parameters:
      - name: "last"
        in: "query"
        required: false
        type: "integer"
      - name: "state"
        in: "query"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Operation"
    post:
      description: "Starts execution of a Multi-Target Application operation\n"
      operationId: "StartMtaOperation"
      parameters:
      - in: "body"
        name: "operation"
        required: true
        schema:
          $ref: "#/definitions/Operation"
      responses:
        202:
          description: "Accepted"
          headers:
            Location:
              type: "string"
              format: "uri"
              description: "A Location header, used for transpassing between different\
                \ resources\n"
  /operations/{operationId}:
    get:
      description: "Retrieves Multi-Target Application operation\n"
      operationId: "GetMtaOperation"
      parameters:
      - name: "operationId"
        in: "path"
        required: true
        type: "string"
      - name: "embed"
        in: "query"
        description: "Adds the specified property in the response body\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Operation"
    post:
      description: "Executes a particular action over Multi-Target Application operation\n"
      operationId: "ExecuteOperationAction"
      parameters:
      - name: "operationId"
        in: "path"
        required: true
        type: "string"
      - name: "actionId"
        in: "query"
        required: true
        type: "string"
      responses:
        202:
          description: "Accepted"
          headers:
            Location:
              type: "string"
              format: "uri"
  /operations/{operationId}/actions:
    get:
      description: "Retrieves available actions for Multi-Target Application operation\n"
      operationId: "GetOperationActions"
      parameters:
      - name: "operationId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              type: "string"
...
securityDefinitions:
  oauth2:
    type: "oauth2"
    tokenUrl: "https://uaa.bosh-lite.com/oauth/token"
    flow: "password"
    scopes: {}
definitions:
...
  Message:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      text:
        type: "string"
      type:
        $ref: "#/definitions/MessageType"
...
  Operation:
    type: "object"
    properties:
      processId:
        type: "string"
      processType:
        type: "string"
      startedAt:
        type: "string"
        format: "date-time"
      endedAt:
        type: "string"
        format: "date-time"
      spaceId:
        type: "string"
      mtaId:
        type: "string"
      namespace:
        type: "string"
      user:
        type: "string"
      acquiredLock:
        type: "boolean"
      cleanedUp:
        type: "boolean"
      state:
        $ref: "#/definitions/State"
      messages:
        type: "array"
        items:
          $ref: "#/definitions/Message"
      parameters:
        type: "object"
        additionalProperties:
          type: "object"
          properties: {}
  State:
    type: "string"
    enum:
    - "RUNNING"
    - "FINISHED"
    - "ERROR"
    - "ABORTED"
    - "ACTION_REQUIRED"
  MessageType:
    type: "string"
    enum:
    - "INFO"
    - "ERROR"
    - "WARNING"
    - "EXT"
    - "TASK_STARTUP"
...